move home page to a component (#44)
* move home page to a component

* fmt

* fix hardcode

* style header and socials image

* new socials banner

* dynamic domain in metatags

* favicon

* preview as testnet domain

* testnet domain config

* formatting

---------
