+++ frontend/constants/constants.ts
@@ -1,4 +1,3 @@
 export const constants = {
-  DOMAIN: "https://burntpix-archives.netlify.app",
   IPFS_GATEWAY: "https://api.universalprofile.cloud/ipfs",
 };
