+++ frontend/constants/networks.ts
@@ -1,4 +1,5 @@
 export interface Network {
+  domain: string;
   chainId: number;
   name: string;
   symbol: string;
@@ -13,6 +14,7 @@ export interface Network {
 
 const NETWORKS = {
   mainnet: {
+    domain: "https://burntpix-archives.netlify.app",
     chainId: 42,
     name: "LUKSO",
     symbol: "LYX",
@@ -26,6 +28,7 @@ const NETWORKS = {
     burntPixArchivesAddress: "0x4b8cCCF15514e136beD7d07365D8FA48353e7947", //TODO update this
   },
   testnet: {
+    domain: "https://testnet--burntpix-archives.netlify.app",
     chainId: 4201,
     name: "LUKSO Testnet",
     symbol: "LYXt",
