+++ frontend/utils/burntPixUtils.ts
@@ -14,15 +14,18 @@ export const refineToArchive = async (
   ).refineToArchive(iters);
 };
 
-export function getNextIterationsGoal(nextLevelIterations: number, currentIterations: number) {
+export function getNextIterationsGoal(
+  nextLevelIterations: number,
+  currentIterations: number,
+) {
   const multiplier = 1000;
   if (nextLevelIterations <= 1) return multiplier;
   let a = 1;
   let b = 1;
   for (let i = 2; i < nextLevelIterations; i++) {
-      const c = a + b;
-      a = b;
-      b = c;
+    const c = a + b;
+    a = b;
+    b = c;
   }
-  return ((a + b) * multiplier) - currentIterations;
+  return (a + b) * multiplier - currentIterations;
 }
